Hotfix 4.9 (Patched all the marvelous bugs i created in hotfix 4.7)
So... In the previous hotfix, i had created a bug where the comboboxes would not reorder themselves after the combobox was changed before editing a patients or clinicians data. thats bug number 1. bug number 2 was, whenever a user editing any part of a clinician or patient THAT WASN'T THEIR PHOTO, the photo of that patient or clinician would be deleted from inside the database. BUG NUMBER 3, if any comboboxes were empty, the program would just shit the bed and break. i hate to admit that this did take me about an hour to patch up, and no one will be able to appreciate it. BUT I WILL!!